Split AE2 integration into facade/interface/stub/impl to isolate all AE2 code and load it only when ae2 mod is present, Removed all AE2 types from shared APIs and replaced with safe wrappers fixed logout crash caused by AE2 class loading on missing dependency and added safe AE2 detection with Class.forName + hard runtime guards, fixed rate limiter edge cases, O(n2) cleanup (now heap-based), null player handling and improved cache/thread safety, hardened the reflective AE2 access and updated AE2 dependency
